src/requirements.txt
Django==2.1.7
django-appconf==1.0.3
django-imagekit==4.0.2
pilkit==2.0
Pillow==6.0.0
psycopg2==2.8.2
python-decouple==3.1
pytz==2018.9
six==1.12.0